refactor(cli): split alert command into submodules [OPE-444]
